support syncing ConfigMap's with config.xml files <-> Jenkins Jobs
for non-pipeline jobs lets support the creation of Jenkins jobs from ConfigMap resources labelled correctly
that store the config.xml in the ConfigMap data.

This lets us create jobs like the github org job via the kubernetes REST API without requiring access to a Jenkins master for openshift.io/fabric8.

One day we could have a new BuildConfig kind in openshift and we could migrate to using that; but until then this simple approach allows us to use ConfigMaps for arbitrary jenkins jobs which are not pipelines