Fix finding buildConfig at the time of webhook push event
Previously we have fixed the issue of finding buildConfig if app name and repo name
are different by finding the buildConfig perfectly. This issue still occurs for
webhook event on master and build is not getting created.

So made a function in openshiftUtils to perfectly find the buildConfig and
used it at both places
Fixes #35
Fixes https://github.com/openshiftio/openshift.io/issues/3969

Refactor according to the PR comments

Add Test

Refactoring